% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bioc_maint_packs}
\alias{bioc_maint_packs}
\title{a data.frame of packages maintained by Bioconductor Package Maintainer on June 11 2021}
\format{
data.frame
}
\usage{
bioc_maint_packs
}
\description{
a data.frame of packages maintained by Bioconductor Package Maintainer on June 11 2021
}
\examples{
head(bioc_maint_packs)
}
\keyword{datasets}
