% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_bccheck_out}
\alias{demo_bccheck_out}
\title{a list of BiocCheck::BiocCheck results for demonstration purposes, produced nov 2020}
\format{
list
}
\usage{
demo_bccheck_out
}
\description{
a list of BiocCheck::BiocCheck results for demonstration purposes, produced nov 2020
}
\examples{
names(demo_bccheck_out)
}
\keyword{datasets}
